feat(security): fix OWASP security vulnerabilities (auth bypass, path traversal, office signatures, missing headers)
